Modified Model weight to float

+++ src/main/java/com/github/onsdigital/utils/nlp/VectorModel.java
@@ -17,6 +17,6 @@ public interface VectorModel {
 
     String getModelFilename();
 
-    double getWeight();
+    float getWeight();
 
 }
